+++ nanos-lite/src/fs.c
@@ -85,7 +85,7 @@ static int GetOpenInd(size_t fd)
 int fs_open(const char *pathname, int flags, int mode)
 {
   //printf("%d\n",LENGTH(file_table));
-  for (size_t i = 0; i < LENGTH(file_table); i++)
+  for (int i = 0; i < LENGTH(file_table); i++)
   {
     //printf("%s %s %d\n",pathname,file_table[i].name,file_table[i].size);
     //if(i>=6) printf("%s %s\n",pathname,file_table[i].name);
