+++ navy-apps/libs/libndl/NDL.c
@@ -77,7 +77,7 @@ void NDL_DrawRect(uint32_t *pixels, int x, int y, int w, int h) {
   //AM_GPU_CONFIG_T ev = io_read(AM_GPU_CONFIG);
   ///uint32_t height = ev.height;
   //uint32_t width = ev.width;
-  for (size_t i = 0; i < h &&(y+i) < Canvas_h; i++)
+  for (int i = 0; i < h &&(y+i) < Canvas_h; i++)
   {
     uint64_t offset = (y+Canvas_y+i)*screen_w+(x+Canvas_x);
     //printf("offset %d screen_w %d\n",offset,screen_w);
