+++ nemu/src/cpu/difftest/ref.c
@@ -26,7 +26,7 @@ void difftest_memcpy(paddr_t addr, void *buf, size_t n, bool direction) {
   
   if (direction == DIFFTEST_TO_REF)
   {
-    for (size_t i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
     {
       paddr_write(addr+i,1,((uint8_t *)buf)[i]);
     }
@@ -38,7 +38,7 @@ void difftest_memcpy(paddr_t addr, void *buf, size_t n, bool direction) {
 void difftest_regcpy(void *dut, bool direction) {
   if(direction == DIFFTEST_TO_REF){
     //nemu作为ref//把npc的寄存器拷贝给nemu的cpu
-    for (size_t i = 0; i < 32; i++)
+    for (int i = 0; i < 32; i++)
     {
       cpu.gpr[i] = ((uint64_t *)dut)[i];
     }
@@ -54,7 +54,7 @@ void difftest_regcpy(void *dut, bool direction) {
     cpu.mip = ((uint64_t *)dut)[38];
   }
   else {
-    for (size_t i = 0; i < 32; i++)
+    for (int i = 0; i < 32; i++)
     {
       ((uint64_t *)dut)[i]= cpu.gpr[i];
     }
