+++ nemu/src/isa/riscv32/difftest/dut.c
@@ -19,7 +19,7 @@
 
 bool isa_difftest_checkregs(CPU_state *ref_r, vaddr_t pc) {
   if (ref_r->pc!=pc) return false;
-  for (size_t i = 0; i < 32; i++)
+  for (int i = 0; i < 32; i++)
   {
     if (ref_r->gpr[i]!=cpu.gpr[i])
     {
