+++ nemu/src/isa/riscv64/reg.c
@@ -25,7 +25,7 @@ const char *regs[] = {
 
 void isa_reg_display() {
   printf ("pc: 0x%08lx \n",cpu.pc);
-  for (size_t i = 0; i < 32; i++)
+  for (int i = 0; i < 32; i++)
   {
     printf ("%s \t  info: 0x%08lx\t \n",regs[i],cpu.gpr[i]); 
   }
@@ -50,7 +50,7 @@ word_t isa_reg_str2val(const char *s, bool *success) {
   sscanf(strtemp,"%d",&index);
   return cpu.gpr[index];
   */
- for (size_t i = 0; i < 32; i++)
+ for (int i = 0; i < 32; i++)
  {
     if (strcmp(s,regs[i]) == 0)
     {
